Tweets Module Clean Up
