+++ tweetme2-web/src/tweets/index.js
@@ -1,8 +1,15 @@
-import {ActionBtn, Tweet, TweetsList, TweetsComponent} from './components'
+import {ActionBtn} from './buttons'
+
+import {TweetsComponent} from './components'
+import {TweetCreate} from './create'
+import {Tweet} from './detail'
+import {TweetsList} from './list'
+
 
 export {
     ActionBtn,
     Tweet,
+    TweetCreate,
+    TweetsComponent,
     TweetsList,
-    TweetsComponent
 }
